+++ sdk/keyvault/azure-security-keyvault-secrets/inc/azure/keyvault/keyvault_secrets.hpp
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+/**
+ * @brief Includes all public headers from Azure Template SDK library.
+ *
+ */
+
+#pragma once
+
+#include "azure/keyvault/secrets/dll_import_export.hpp"
+#include "azure/keyvault/secrets/secret_client.hpp"